+++ src/lib/DatabaseController.ts
@@ -1,8 +1,6 @@
 import { DataTypes, Sequelize } from "sequelize";
 import { validateEnvVars, type DbTestResult } from "./db_utils";
-import { EncounterType } from "@/models/enums/EncounterType";
-import { Rarity } from "@/models/enums/Rarity";
-import { Role } from "@/models/enums";
+import type { GuildMemberMapping } from "@/models/types/GuildMemberMapping";
 
 export class DatabaseController {
     private sequelize: Sequelize;
@@ -38,208 +36,22 @@ export class DatabaseController {
             process.exit(1);
         }
 
-        // Guild table
-        const guild = this.sequelize.define("Guilds", {
-            guildId: {
-                type: DataTypes.STRING,
-                allowNull: false,
-                primaryKey: true,
-            },
-            guildTag: {
-                type: DataTypes.STRING,
-                allowNull: false,
-            },
-            name: {
-                type: DataTypes.STRING,
-                allowNull: false,
-            },
-            level: {
-                type: DataTypes.INTEGER,
-                allowNull: false,
-            },
-        });
-
         // GuildMember table
-        const guildMember = this.sequelize.define(
-            "GuildMembers",
-            {
-                userId: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                    primaryKey: true,
-                },
-                role: {
-                    type: DataTypes.ENUM(...Object.values(Role)),
-                    allowNull: false,
-                },
-                level: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                lastActivityOn: {
-                    type: DataTypes.DATE,
-                    allowNull: true,
-                },
-                guildId: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                    references: {
-                        model: "Guilds",
-                        key: "guildId",
-                    },
-                },
-            },
-            {
-                indexes: [
-                    {
-                        fields: ["guildId"],
-                    },
-                ],
-            }
-        );
-
-        // GuldRaidSeason table
-        const guildRaidSeason = this.sequelize.define("GuildRaidSeasons", {
-            guildRaidSeasonId: {
-                type: DataTypes.UUID,
-                defaultValue: DataTypes.UUIDV4,
+        this.sequelize.define("GuildMembers", {
+            userId: {
+                type: DataTypes.STRING,
+                allowNull: false,
                 primaryKey: true,
             },
-            season: {
-                type: DataTypes.INTEGER,
+            username: {
+                type: DataTypes.STRING,
                 allowNull: false,
             },
             guildId: {
                 type: DataTypes.STRING,
                 allowNull: false,
-                references: {
-                    model: "Guilds",
-                    key: "guildId",
-                },
-            },
-            seasonConfigId: {
-                type: DataTypes.STRING,
-                allowNull: false,
             },
         });
-
-        // Raid table
-        const raid = this.sequelize.define(
-            "Raids",
-            {
-                id: {
-                    type: DataTypes.UUID,
-                    defaultValue: DataTypes.UUIDV4,
-                    primaryKey: true,
-                },
-                userId: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                },
-                tier: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                set: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                encounterIndex: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                remainingHp: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                maxHp: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                encounterType: {
-                    type: DataTypes.ENUM(...Object.values(EncounterType)),
-                    allowNull: false,
-                },
-                unitId: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                },
-                type: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                },
-                rarity: {
-                    type: DataTypes.ENUM(...Object.values(Rarity)),
-                    allowNull: false,
-                },
-                damageDealt: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                damageType: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                },
-                startedOn: {
-                    type: DataTypes.DATE,
-                    allowNull: false,
-                },
-                completedOn: {
-                    type: DataTypes.DATE,
-                    allowNull: false,
-                },
-                globalConfigHash: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                },
-                seasonId: {
-                    type: DataTypes.UUID,
-                    allowNull: false,
-                    references: {
-                        model: "GuildRaidSeasons",
-                        key: "guildRaidSeasonId",
-                    },
-                },
-            },
-            {
-                indexes: [
-                    {
-                        fields: ["seasonId"],
-                    },
-                ],
-            }
-        );
-
-        // PublicHeroDetail table - This table is used to store hero details (toons)
-        const publicHeroDetail = this.sequelize.define(
-            "PublicHeroDetail",
-            {
-                unitId: {
-                    type: DataTypes.STRING,
-                    allowNull: false,
-                    primaryKey: true,
-                },
-                power: {
-                    type: DataTypes.INTEGER,
-                    allowNull: false,
-                },
-                raidId: {
-                    type: DataTypes.UUID,
-                    allowNull: false,
-                    references: {
-                        model: "Raids",
-                        key: "id",
-                    },
-                },
-            },
-            {
-                indexes: [
-                    {
-                        fields: ["raidId"],
-                    },
-                ],
-            }
-        );
 
         // User-guild api token table - This table is used to store discord user-guild api token mapping
         this.sequelize.define("discordApiTokenMappings", {
@@ -258,19 +70,14 @@ export class DatabaseController {
         // SCHEMA RELATIONSHIPS
         //  */
         // Define relationships with explicit foreign keys
-        guild.hasMany(guildMember, { foreignKey: "guildId" });
-        guildMember.belongsTo(guild, { foreignKey: "guildId" });
 
-        guildRaidSeason.hasMany(raid, { foreignKey: "seasonId" });
-        raid.belongsTo(guildRaidSeason, { foreignKey: "seasonId" });
-
-        guild.hasMany(guildRaidSeason, { foreignKey: "guildId" });
-        guildRaidSeason.belongsTo(guild, { foreignKey: "guildId" });
-
-        raid.hasMany(publicHeroDetail, { foreignKey: "raidId" });
-        publicHeroDetail.belongsTo(raid, { foreignKey: "raidId" });
-
-        this.sequelize.sync({});
+        try {
+            await this.sequelize.sync({ force: false });
+            console.log("Models defined and database synced successfully.");
+        } catch (error) {
+            console.error("Error syncing database:", error);
+            process.exit(1);
+        }
     }
 
     public async isReady(): Promise<DbTestResult> {
@@ -319,6 +126,109 @@ export class DatabaseController {
         } catch (error) {
             console.error("Error retrieving user token from database:", error);
             return null;
+        }
+    }
+
+    public async getGuildMembersByGuildId(guildId: string) {
+        try {
+            const result = await this.sequelize.models["GuildMembers"]?.findAll(
+                {
+                    where: {
+                        guildId: guildId,
+                    },
+                }
+            );
+            if (!result) {
+                return null;
+            }
+
+            return result.map((member) => {
+                return {
+                    userId: member.getDataValue("userId") as string,
+                    username: member.getDataValue("username") as string,
+                } as GuildMemberMapping;
+            });
+        } catch (error) {
+            console.error("Error retrieving player name from database:", error);
+            return null;
+        }
+    }
+    public async getPlayerName(userId: string): Promise<string | null> {
+        try {
+            const result = await this.sequelize.models["GuildMembers"]?.findOne(
+                {
+                    where: {
+                        userId: userId,
+                    },
+                }
+            );
+            if (!result) {
+                return null;
+            }
+
+            return result.getDataValue("username") as string;
+        } catch (error) {
+            console.error("Error retrieving player name from database:", error);
+            return null;
+        }
+    }
+
+    public async updatePlayerName(
+        userId: string,
+        name: string,
+        guildId: string
+    ): Promise<boolean> {
+        try {
+            const guildMembersModel = this.sequelize.models["GuildMembers"];
+            if (!guildMembersModel) {
+                console.error("GuildMembers model is not defined.");
+                return false;
+            }
+
+            const res = await guildMembersModel.upsert({
+                userId: userId,
+                username: name,
+                guildId: guildId,
+            });
+
+            if (!res) {
+                console.error("Error updating player name in database.");
+                return false;
+            }
+
+            return true;
+        } catch (error) {
+            console.error(
+                `Error updating player name for userId: ${userId}, name: ${name}`,
+                error
+            );
+            return false;
+        }
+    }
+
+    public async deletePlayerName(
+        userId: string,
+        guildId: string
+    ): Promise<number> {
+        try {
+            const guildMembersModel = this.sequelize.models["GuildMembers"];
+            if (!guildMembersModel) {
+                console.error("GuildMembers model is not defined.");
+                return -1;
+            }
+
+            return await guildMembersModel.destroy({
+                where: {
+                    userId: userId,
+                    guildId: guildId,
+                },
+            });
+        } catch (error) {
+            console.error(
+                `Error deleting player name for userId: ${userId}`,
+                error
+            );
+            return -1;
         }
     }
 }