+++ src/lib/db_utils.ts
@@ -1,5 +1,3 @@
-import { Sequelize } from "sequelize";
-
 export interface DbTestResult {
     isSuccess: boolean;
     message?: string;