+++ src/models/types/GuildMemberMapping.ts
@@ -1,0 +1,4 @@
+export interface GuildMemberMapping {
+    userId: string;
+    username: string;
+}